+++ packages/ice/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v3.0.6
+
+- [fix] remove code when dead code is `ObjectProperties`
+- [fix] rebuild server entry when document changed
+
+## v3.0.5
+
+- [fix] bump @swc/helpers version (0.4.12-> 0.4.14) for module cannot resolve of `@swc/helpers/src/_object_destructuring_empty.mjs`
+
 ## v3.0.4
 
 - [feat] support incremental compile for server bundle